:truck: Move ensure_default_language to translation module
